Modifying the start function in ImageJob. Definition will be be placed outside OpenSeadragon.
Ok, introducing the "startFunction" is not the proper
way to implement loading tiles from
single files. TODO: Redesign this.
Right now startFunction is meant to be defined
outside OpenSeadragon, and serves as a way to modify
the behaviour of OpenSeadragon without having to have the code
inside the OpenSeadragon file tree.

Some of the code in the start function was moved into
the new function setupImageCallbacks().